PLF-8098 : update Elasticsearch from 5.6.3 to 5.6.11
Upgrade Elasticsearch from 5.6.3 to 5.6.11.
Elasticsearch 5.6.11 comes with some dependencies upgrades, including a Tika upgrade from 1.14 to 1.18. In this new version, a warning is logged when a parser is missing, so the new dependencies org.apache.pdfbox:jbig2-imageio and com.github.jai-imageio:jai-imageio-jpeg2000 have been added to ingest-attachment plugin dependencies.